rename eventcallback to simple ts
