+++ js/renovation/ui/pager/common/keyboard_action_context.tsx
@@ -1,6 +1,6 @@
 import { createContext } from '@devextreme-generator/declarations';
 import { DisposeEffectReturn } from '../../../utils/effect_return.d';
-import { EventCallback } from '../../common/event_callback.d';
+import { EventCallback } from '../../common/event_callback';
 
 export interface KeyboardActionContextType {
   registerKeyboardAction: (el: HTMLElement, handler: EventCallback) => DisposeEffectReturn;
