+++ js/renovation/ui/pager/common/pager_props.ts
@@ -1,7 +1,7 @@
 import {
   ComponentBindings, OneWay, TwoWay, Event,
 } from '@devextreme-generator/declarations';
-import { EventCallback } from '../../common/event_callback.d';
+import { EventCallback } from '../../common/event_callback';
 
 export type DisplayMode = 'adaptive' | 'compact' | 'full';
