+++ js/renovation/ui/pager/content.tsx
@@ -15,7 +15,7 @@ import { combineClasses } from '../../utils/combine_classes';
 import { Widget } from '../common/widget';
 import { DisposeEffectReturn } from '../../utils/effect_return.d';
 import { registerKeyboardAction } from '../../../ui/shared/accessibility';
-import { EventCallback } from '../common/event_callback.d';
+import { EventCallback } from '../common/event_callback';
 import { KeyboardActionContext, KeyboardActionContextType } from './common/keyboard_action_context';
 
 export const viewFunction = ({
